stream/requirements.txt
ffmpeg-python
Pillow
qrcode
colorthief
requests
numpy